src/main/kotlin/snowesamosc/kotlinjsonparser/factory/JsonLiteralImpl.kt
package snowesamosc.kotlinjsonparser.factory

import snowesamosc.kotlinjsonparser.node.JsonNode

internal sealed class JsonLiteralImpl(
    private val originalString: String,
    private val children: List<JsonLiteral>
) : JsonLiteral {

    override fun asString(): String = originalString

    override fun hasChildren(): Boolean = children.isNotEmpty()

    override fun getChildren(): List<JsonLiteral> = children

    override fun isJsonNode(): Boolean = false

    override fun asJsonNode(): JsonNode? = null

    override fun toString(): String {
        return getName() + if (hasChildren()) {
            getChildren().toString()
        } else {
            ""
        }
    }

    abstract fun getName(): String

    internal class WS private constructor(
        originalString: String
    ) : JsonLiteralImpl(originalString, emptyList()) {
        override fun getName(): String = "WS"

        companion object Factory {
            fun greedyCreate(str: String): GreedyCreateResult {
                val nodeBuilder = StringBuilder()
                val originalStringBuilder = StringBuilder(str)

                CharLoop@
                for (c in str) {
                    when (c) {
                        '\u0020', '\u0009', '\u000A', '\u000D' -> {
                            nodeBuilder.append(c)
                            originalStringBuilder.deleteCharAt(0)
                        }
                        else -> break@CharLoop
                    }
                }

                val ws = WS(nodeBuilder.toString())

                return GreedyCreateResult(originalStringBuilder.toString(), ws)
            }
        }
    }
}